refactor: move update notifications read statuses
Move the function to update the notification's read statuses into the
NotificationProvider to be used by other components through the context.
